+++ system/scripts/items/advancement-sheet.js
@@ -14,9 +14,6 @@ export class AdvancementSheetL5r5e extends ItemSheetL5r5e {
         return foundry.utils.mergeObject(super.defaultOptions, {
             classes: ["l5r5e", "sheet", "advancement"],
             template: CONFIG.l5r5e.paths.templates + "items/advancement/advancement-sheet.html",
-            width: 520,
-            height: 480,
-            tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }],
         });
     }