+++ system/scripts/items/item-sheet.js
@@ -10,9 +10,6 @@ export class ItemSheetL5r5e extends BaseItemSheetL5r5e {
         return foundry.utils.mergeObject(super.defaultOptions, {
             classes: ["l5r5e", "sheet", "item"],
             template: CONFIG.l5r5e.paths.templates + "items/item/item-sheet.html",
-            width: 520,
-            height: 480,
-            tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }],
         });
     }