+++ system/scripts/items/weapon-sheet.js
@@ -9,9 +9,6 @@ export class WeaponSheetL5r5e extends ItemSheetL5r5e {
         return foundry.utils.mergeObject(super.defaultOptions, {
             classes: ["l5r5e", "sheet", "weapon"],
             template: CONFIG.l5r5e.paths.templates + "items/weapon/weapon-sheet.html",
-            width: 520,
-            height: 480,
-            tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }],
         });
     }